feat: add dynamic metadata for search results (#842)
<!-- 👋 Hi, thanks for sending a PR to tidelift-me-up-site! 💖.
Please fill out all fields below and make sure each item is true and [x]
checked.
Otherwise we may not be able to review your PR. -->

## PR Checklist

- [x] Addresses an existing open issue: fixes #67 
- [x] That issue was marked as [`status: accepting
prs`](https://github.com/JoshuaKGoldberg/tidelift-me-up-site/issues?q=is%3Aopen+is%3Aissue+label%3A%22status%3A+accepting+prs%22)
- [x] Steps in
[CONTRIBUTING.md](https://github.com/JoshuaKGoldberg/tidelift-me-up-site/blob/main/.github/CONTRIBUTING.md)
were taken

## Overview

<!-- Description of what is changed and how the code change does that.
-->

This PR introduces dynamic metadata generation for the home/search page.
It's metadata is now dynamically generated based on the `searchParams`
in the URL:

> {username} | Tidelift Me Up
> {username} has {packageCount} npm packages eligible for Tidelift
funding. 💸

This includes when a valid user has 0 packages eligible for funding
based on search parameters. The metadata will look like the following:

> {username} | Tidelift Me Up
> {username} has 0 npm packages eligible for Tidelift funding. 💸

When no `searchParams` are provided, the default metadata is used, which
is defined in `layout.tsx`:

> Tidelift Me Up
> Find your npm packages eligible for Tidelift funding. 💸

Also, something to note is that the API-related logic that was in
`page.tsx` has been extracted into a separate file to improve code
readability and maintainability.